refactor(openai): changed to brainpicking
